Замена алгоритма Бридсона на обычный рандом
Нужны тесты
